Ysh (#13)
* 세팅 수정

* 상품 추가 기능 구현

* 상품 구맹 기능 구현

* 상품 수량 0개식 자동 삭제

* 로그인

* 로그인 기능 구현

* 삭제 기능 구현

* 업데이트 기능 구현

* 상품 수정 기능 구현

* list 배치 css 수정

* 상품정보 덥데이트 숮어

(cherry picked from commit dbcdeb1ddd439e7e186de153a9fb19aea5e507a6)

+++ src/main/java/com/example/ProductController.java
@@ -4,23 +4,21 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.util.StringUtils;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 @Controller
+@RequestMapping("/products")
 public class ProductController {
     @Autowired
     ProductDao productDao;
     @Autowired
     LoginInterceptor loginInterceptor;
 
-    @GetMapping("/products")
+    @GetMapping
     public String getProducts(@RequestParam(required = false) String keyword, Model model) {
         List<ProductVO> products;
 
@@ -35,15 +33,17 @@ public class ProductController {
         return "list";
     }
 
-    @GetMapping("/products/{itemId}")
+    public boolean checkLogin(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        return loginInterceptor.preHandle(request, response, null);
+    }
+
+    @GetMapping("/{itemId}")
     public String getProductDetails(@PathVariable Integer itemId, Model model, HttpServletRequest request, HttpServletResponse response) {
         try {
-            if (!loginInterceptor.preHandle(request, response, null)) {
-                // 세션이 유효하지 않은 경우 로그인 페이지로 리다이렉트
+            if (!checkLogin(request, response)) {
                 return "redirect:/login/login";
             }
 
-            // 세션이 유효한 경우에만 실행되는 코드
             ProductVO product = productDao.getProduct(itemId);
             model.addAttribute("product", product);
             return "view";
@@ -52,14 +52,16 @@ public class ProductController {
         }
     }
 
-    @GetMapping("/products/add")
+    @GetMapping("/add")
     public String addProducts(HttpServletRequest request, HttpServletResponse response) {
         try {
             if (!loginInterceptor.preHandle(request, response, null)) {
+                System.out.println("로그인 안됨");
                 // 세션이 유효하지 않은 경우 로그인 페이지로 리다이렉트
                 return "redirect:/login/login";
             }
 
+            System.out.println("로그인 됨");
             // 세션이 유효한 경우에만 실행되는 코드
             return "add";
         } catch (Exception e) {
@@ -67,13 +69,13 @@ public class ProductController {
         }
     }
 
-    @PostMapping("/products/addOk")
+    @PostMapping("/addOk")
     public String addProducts(ProductVO productVO) {
         productDao.insertProduct(productVO);
         return "redirect:/products";
     }
 
-    @GetMapping("/products/edit")
+    @GetMapping("/edit")
     public String updateProducts(HttpServletRequest request, HttpServletResponse response) {
         try {
             if (!loginInterceptor.preHandle(request, response, null)) {
@@ -88,27 +90,55 @@ public class ProductController {
         }
     }
 
-    @PostMapping("/products/editOk")
-    public String updateProducts(ProductVO productVO) {
-        productDao.insertProduct(productVO);
-        return "redirect:/products";
-    }
-
-    @PostMapping("/products/update/{id}")
+    @PostMapping("/update/{id}")
     public String updateProducts(@PathVariable Integer id) {
         ProductVO productVO = productDao.getProduct(id);
         productDao.purchaseProduct(productVO);
         return "redirect:/products";
     }
 
-    @GetMapping("/products/delete/{id}")
-    public String deletePost(@PathVariable int id) {
-        if (productDao.deleteProduct(id) != 0) {
-            System.out.println("success");
-        } else {
-            System.out.println("fail");
-        }
+    @GetMapping("/edit/{id}")
+    public String editPost(@PathVariable int id, Model model, HttpServletRequest request, HttpServletResponse response) {
+        try {
+            if (!loginInterceptor.preHandle(request, response, null)) {
+                // 세션이 유효하지 않은 경우 로그인 페이지로 리다이렉트
+                return "redirect:/login/login";
+            }
 
-        return "redirect:../../products";
+            // 세션이 유효한 경우에만 실행되는 코드
+            ProductVO product = productDao.getProduct(id);
+            model.addAttribute("product", product);
+            return "edit";
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
+
+    @PostMapping("/editOk")
+    public String updateProducts(ProductVO productVO) {
+        productDao.updateProduct(productVO);
+        return "redirect:/products";
+    }
+
+    @GetMapping("/delete/{id}")
+    public String deletePost(@PathVariable int id, HttpServletRequest request, HttpServletResponse response) {
+        try {
+            if (!loginInterceptor.preHandle(request, response, null)) {
+                // 세션이 유효하지 않은 경우 로그인 페이지로 리다이렉트
+                return "redirect:/login/login";
+            }
+
+            // 세션이 유효한 경우에만 실행되는 코드
+            if (productDao.deleteProduct(id) != 0) {
+                System.out.println("success");
+            } else {
+                System.out.println("fail");
+            }
+
+            return "redirect:../../products";
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
 }
